fix rt75680 - when working with utf8 strings one must always use s+=UTF8SKIP(s) to move to the next char
Most of the regex code where do the two types of increments are wrapped up in macros.

Unfortunately these macros arent suitable in this case because we use goto to jump
inside the loop under some situations, and since this is a one-off case I figured the
modest C&P associated was better than creating a new macro just for this case.

There is still a possible bug here marked by an XXX, which will need to be fixed
once I find out the correct way to simulate strptr--. Additionally I havent found
a test case that actually exposes this form of the bug.

Moral of the story, utf8 makes string scanning awkward... And slow...