mktables: Silence common build warning
The first time mktables is run on a clean directory, mktables.lst is
non-existent.  Rather than warn on this common occurrence, just assume
it is ok, and continue.  The point of the warning was to point out a
potential problem.  But it occurs so often that people have gotten
inured (habituated) to it, and it has lost all effectiveness.  Better to
not output it, and have any problems show up as repeated rebuilds that
will annoy enough to be dealt with.