autodoc.pl: Clarify use of undocumented fcns
perlapi.pod was changed to list all the undocumented functions.  The
introductory text was not updated at that time.  Now every public API
function is at least listed in this pod.  Clarify that functions not
listed are not public and should not be used, and add a plea for
documenting the ones which are missing theirs.