perlunicode.pod: Fix misleading info, expand
There was some misleading, or uncharitably, wrong text in this pod about
user-defined casing.  And, it jumped the gun, presuming that 5.14 would
fix something for which there has not been a patch submitted yet.

And, I realized there was a way around having to figure out the utf8 for
a character.