In mktables, disable string overloading when de-duping lists.
As the unique() function is called with either a list of strings or a list of
objects, and all objects stringify to distinct values, there is no behaviour
difference between using object stringification and object address for a
"seen" hash. There is, however, an 8% run time difference :-)

Devel::NYTProf strikes again.