regcomp.c: Use longjmp to abandon first pass quickly
regcomp.c currently does a first pass to determine the size needed for
the regex.  If the regex needs to be in utf8, but this wasn't known at
the beginning of the pass the size computation needs to be completely
redone with that in mind.  The input is converted to utf8 and the first
pass is redone.  Prior to this patch, the discovery of needing to redo
the first pass merely set a flag and continued the now useless rest of
the first pass.  This patch causes this first pass to be aborted
immediately upon discovering that it will have to be redone.

This uses the Perl macros that wrap setjmp and longjmp.  When the first
pass is abandoned, it longjmps back to the corresponding setjmp to
convert to utf8 and redo.  I was advised that I could use setjmp and
longjump directly, but it seemed safer to use the wrappers, as they
should know about any platform-dependent issues.

If this code intercepts a longjmp that wasn't meant for it, it simply
reissues the longjmp so that the correct handler will get it.

This patch should have no effect on externally visible behavior, except
for a speedup of some regex compiles.