Change the first argument of Perl_fetch_cop_label() to COP *
From a suggestion from Ben Morrow.

The first argument used to be struct refcounted_he *, which exposed an
implementation detail - that the COP's labels are (now) stored in this way.

Google Code Search and an unpacked CPAN both fail to find any users of this
API, so the impact should be minimal.