Back out the mauve module and related changes
It's was intended as a temporary namespace only, and we really don't want to
ship it in any release until we've figured out what it should really look like.

This reverts commit 05c0d6bbe3ec5cc9af99d105b8648ad02ed7cc95,
  "add sv_reftype_len() and make sv_reftype() be a wrapper for it"
commit 792477b9c2e4c75cb03d07bd6d25dc7e1fdf448e,
 "create the "mauve" temporary namespace for things like reftype"
commit 8df6b97c1de8326d50ac9c8cae4bf716393b45bb,
  "mauve.t needs access to %Config, make sure it's available"
commit cfe9162d0d593cd12a979c73df82c7509b324343,
 "use more efficient sv_reftype_len() interface"
and commit 47b13905e23c2a72acdde8bb4669e25e5eaefec4
  "add more tests to lib/mauve.t so it tests also that mauve::reftype can return "LVALUE""

There's a `mauve' branch still containing all the code for the temporary mauve
namespace. That should be used to work on it until it's mostly ready to be
released, and only then merged to blead. Alternatively, it should be deleted if
another way to provide mauve's features in the core is found.