perlfunc: list each/keys/values as ARRAY functions
As of Perl 5.12.0, the each(), keys() and values() functions also
operate on arrays, but were not listed in the table of contents section
titled "Functions for real @ARRAYs".  This patch fixes that oversight.

The delete() and exists() functions were not added, as their use on
array elements has been deprecated.