Remove obsolete core test directory boilerplate from bignum.
This was only needed for testing in the core, when the core's tests all ran in
the top level t/ directory. Without this getting in the way, we don't need
t/TEST and t/harness to run the tests with absolute paths in @INC. Testing in
the CPAN distribution is unaffected.