Teach Perl about Unicode named character sequences
mktables is changed to process the Unicode named sequence file.
charnames.pm is changed to cache the looked-up values in utf8.  A new
function, string_vianame is created that can handle named sequences, as
the interface for vianame cannot.  The subroutine lookup_name() is
slightly refactored to do almost all of the common work for \N{} and the
vianame routines.  It now understands named sequences as created my
mktables..

tests and documentation are added.  In the randomized testing section,
half use vianame() and half string_vianame().