make sure perl doesn't touch stat.t between stats
ebcfa0534 changed stat.t to avoid using TEST as the stat target, since
it is read by other test scripts.  But with the initial stat() in the
BEGIN block perl is still reading the script for compilation, which
will update the access time.  Since the system clock usually didn't
tick another second between the first and second stats, the test
usually succeeded, but occasionally the clock would tick, and the test
would fail.

Moved the stat() out of the BEGIN block to avoid that.

As a check I temporarily added a 2 second sleep after the initial
stat to ensure we didn't have a similar problem to that will
allowed this to pass most of the time.