handy.h: Change isFOO_A() to be O(1) performance
This patch changes the macros whose names end in _A to use table lookup
except for the one (isASCII) which always has only one comparison.

The table is in l1_char_class_tab.h.

The advantage of this is speed.  It replaces some fairly complicated
expressions with an O(1) look-up and a mask.

It uses the FITS_IN_8_BITS() macro to guarantee that the table bounds
are not exceeded.  For legal inputs that are byte size, the optimizer
should get rid of this macro leaving only the lookup and mask.

(This commit was changed from its original form by Steffen.)