The rest of [perl #71710] fixes for File::Find
The code is from Alex Davies. The commit message is from the committer.

Somehow I only ended up with half the patch applied. No tests failed
for me, because these code paths are only reached on Windows, which,
thankfully, I am not using.

Hopefully this will eliminate the smoke that is already pouring in....

(It’s *my* turn to break blead for Windows. :-)