handy.h: Fix so x2p compiles
The recent series of commits on handy.h causes x2p to not compile.
These commits had some differences from what I submitted, in that they
moved the new table to a new header file instead of the submitted
perl.h.   Unfortunately, this bypasses code in perl.h that figures
out about duplicate definitions, and externs, and so fails on programs
that include handy.h but not perl.h.

This patch changes things so that the table lookup is not used unless
perl.h is included.  This is essentially my original patch, but adding
an #include of the new header file.